non_fnv/requirements.txt
absl-py==0.12.0
antlr4-python3-runtime==4.8
anyio==3.4.0
apache-beam==2.34.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==21.2.0
avro-python3==1.9.2.1
Babel==2.9.1
backcall==0.2.0
black==21.4b2
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
contextlib2==21.6.0
crcmod==1.7
cycler==0.11.0
Cython==0.29.26
dataclasses==0.6
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@6999554d5a02e0c9c7a22c1abd8f6eb6812fad0a
dill==0.3.1.1
dm-tree==0.1.6
docopt==0.6.2
easyocr==1.4.1
entrypoints==0.3
et-xmlfile==1.1.0
fastavro==1.4.7
fire==0.4.0
flatbuffers==2.0
fonttools==4.28.5
future==0.18.2
fuzzywuzzy==0.18.0
fvcore==0.1.5.post20220119
gast==0.4.0
gin-config==0.5.0
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-vision==2.6.3
google-pasta==0.2.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
h5py==3.6.0
hdfs==2.6.0
httplib2==0.19.1
hydra-core==1.1.1
idna==3.3
imageio==2.9.0
imagesize==1.3.0
imantics==0.1.12
imgaug==0.4.0
importlib-metadata==4.10.0
imutils==0.5.4
iopath==0.1.9
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.3.2
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.13.1
jupyterlab==3.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.1
kaggle==1.5.12
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
labelImg==1.8.6
labelme2coco==0.2.1
libclang==12.0.0
lvis==0.5.3
lxml==4.7.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
nltk==3.7
notebook==6.4.6
numpy==1.20.3
oauth2client==4.1.3
oauthlib==3.1.1
object-detection==0.1
omegaconf==2.1.1
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
openpyxl==3.0.9
opt-einsum==3.3.0
orjson==3.6.5
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pixellib==0.7.1
platformdirs==2.4.1
plotly==5.5.0
portalocker==2.3.2
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.24
proto-plus==1.19.8
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
py-cpuinfo==8.0.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.3
pycparser==2.21
pydot==1.4.2
Pygments==2.10.0
pymongo==3.12.3
pyparsing==2.4.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.0
pytesseract==0.3.8
python-bidi==0.4.2
python-dateutil==2.8.2
python-Levenshtein==0.12.2
python-slugify==5.0.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
sacrebleu==2.0.0
sahi==0.8.22
scikit-image==0.19.1
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
sentencepiece==0.1.96
seqeval==1.2.2
Shapely==1.8.0
six==1.16.0
sniffio==1.2.0
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-addons==0.15.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io==0.23.1
tensorflow-io-gcs-filesystem==0.23.1
tensorflow-metadata==1.5.0
tensorflow-model-optimization==0.7.0
tensorflow-text==2.7.3
termcolor==1.1.0
terminado==0.12.1
terminaltables==3.1.10
testpath==0.5.0
text-unidecode==1.3
tf-models-official==2.7.0
tf-slim==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
tk==0.1.0
toml==0.10.2
tomli==2.0.0
torch==1.10.1
torchvision==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typeguard==2.13.3
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
wrapt==1.13.3
xmljson==0.2.1
yacs==0.1.8
zipp==3.6.0
